Feat/AE-70 neurad wod

* fix: output_folder from wod dataparser same as data path

* Merge branch 'doc/git-hub-readme' into feat/AE-70-neurad-wod

* clean: typos

* clean: version of dataset used.

* doc: add image resolution info for metrics

* Merge remote-tracking branch 'origin/feat/AE-70-neurad-wod' into doc/git-hub-readme

* Merged in doc/git-hub-readme (pull request #3)

doc: add image resolution info for metrics

* update: merged main branch from github back prior pull request

* doc: commented torch dynamo suppress error message

* revert: reverted changes to neurad config steps

* revert: reverted changes to neurad config steps on viewer side

* fix: removed wod api from requirements and put back to readme because of too strict requirements

* feat: added support for waymo-open-dataset

* revert: change to .vscode/launch.json

* revert: change to settings.json

* fix: typo same variable two times in function signature

* fix tiny caracter

* Merge branch 'wod-dataparser' into feat/AE-70-neurad-wod
Merge back minor changes from github

* revert: vscode/*.json files

* fix: typo in render.py

* feature: added support for waymo-open-dataset

* doc: added contributors to waymo readme

* doc: explain waymo-open-dataset dependency installation

* fix: added waymo-open-dataset package installation in Dockerfile

* clean last changes from github


Approved-by: Stanguennec, Julien

+++ nerfstudio/data/dataparsers/base_dataparser.py
@@ -68,7 +68,7 @@ class DataparserOutputs:
     """
     dataparser_transform: Float[Tensor, "3 4"] = torch.eye(4)[:3, :]
     """Transform applied by the dataparser to the entire scene."""
-    actor_transform: Float[Tensor, "3 4"] = torch.eye(4)[:3, :]
+    actor_transform: Float[Tensor, "3 4"] = torch.eye(4, dtype=torch.float32)[:3, :]
     """Transform applied by the dataparser to each actor's local frame."""
     dataparser_scale: float = 1.0
     """Scale applied by the dataparser."""
